+++ src/kel/media/__init__.py
@@ -7,6 +7,7 @@ from kel.media.registry import (
     get_video_model,
     register_media_provider,
 )
+from kel.media.replicate import ReplicateJobHandle, ReplicateMediaModel
 from kel.media.types import MediaResult
 
 __all__ = [
@@ -15,6 +16,8 @@ __all__ = [
     "FalSTTProvider",
     "FalTTSProvider",
     "MediaResult",
+    "ReplicateJobHandle",
+    "ReplicateMediaModel",
     "get_audio_model",
     "get_image_model",
     "get_lipsync_model",